+++ src/test/java/com/qcloud/cos/SuiteTest.java
@@ -2,6 +2,7 @@ package com.qcloud.cos;
 
 import com.qcloud.cos.utils.CRC64Test;
 import com.qcloud.cos.utils.UrlEncoderUtilsTest;
+
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
@@ -9,7 +10,7 @@ import org.junit.runners.Suite.SuiteClasses;
 @RunWith(Suite.class)
 @SuiteClasses({AclTest.class, BatchDeleteTest.class, BucketReplicationTest.class,
         BucketVersioningTest.class, CORSTest.class, CreateDeleteHeadBucketTest.class,
-        GeneratePresignedUrlTest.class, GetBucketLocationTest.class, GetServiceTest.class,
+        GeneratePresignedUrlTest.class, GetBucketLocationTest.class,
         ListObjectTest.class, ListVersionsTest.class, MultipartUploadTest.class,
         PutGetDelTest.class, PutGetLifeCycleConfigTest.class, PutObjectCopyTest.class, 
         SSECustomerTest.class, TransferManagerTest.class, BucketWebsiteTest.class, AppendObjectTest.class,